Remove both Google Vertex & AWS SageMaker from `Dockerfile-cuda-all`
Neither of those is required, since both Google Cloud and AWS SageMaker
port the `Dockerfile-cuda-all` as-is, and then re-builds it there,
meaning that the actual BUILD_ARG for VERTEX is not being used at all,
neither the AWS SageMaker stage. For more information check the
repositories
https://github.com/huggingface/Google-Cloud-Containers/tree/main/containers/tei,
and
https://github.com/awslabs/llm-hosting-container/tree/main/huggingface/pytorch/tei/docker,
respectively for Google Cloud and AWS SageMaker.