Fix `cargo install` commands for `candle` with CUDA
The `cargo install --path router ...` commands for the features
`candle-cuda` and `candle-cuda-turing` were using the flag
`--no-default-features` that was ignoring the default features in the
`text-embeddings-router` being `candle,http,dynamic-linking`, hence the
command was failing as `cudarc` requires to have any linking out of
`static-linking`, `dynamic-linking` or `dynamic-loading` for the CUDA
Libraries, hence when ignoring the `default` features, and just adding
`-F http` either the e.g. `-F dynamic-loading` feature was missing or
just to only keep the e.g. `-F candle-cuda`, as `dynamic-linking` and
`http` are default features already.